rtcstats: move ws connection to external module
and pass trace function
