+++ nonmodule.js
@@ -1,6 +1,6 @@
 // helper script to create the non-module version with some defaults.
 require('./rtcstats')(
-    'wss://rtcstats.tokbox.com',
+    require('./trace-ws')('wss://rtcstats.tokbox.com'),
     1000,
     ['', 'webkit', 'moz']
 );
