+++ rtcstats.js
@@ -108,44 +108,7 @@ function removeTimestamps(results) {
 }
 */
 
-module.exports = function(wsURL, getStatsInterval, prefixesToWrap) {
-  var PROTOCOL_VERSION = '1.0';
-  var buffer = [];
-  var connection = new WebSocket(wsURL + window.location.pathname, PROTOCOL_VERSION);
-  connection.onerror = function(e) {
-    console.log('WS ERROR', e);
-  };
-
-  /*
-  connection.onclose = function() {
-    // reconnect?
-  };
-  */
-
-  connection.onopen = function() {
-    while (buffer.length) {
-      connection.send(JSON.stringify(buffer.shift()));
-    }
-  };
-
-  /*
-  connection.onmessage = function(msg) {
-    // no messages from the server defined yet.
-  };
-  */
-
-  function trace() {
-    //console.log.apply(console, arguments);
-    // TODO: drop getStats when not connected?
-    var args = Array.prototype.slice.call(arguments);
-    args.push(new Date().getTime());
-    if (connection.readyState === 1) {
-      connection.send(JSON.stringify(args));
-    } else {
-      buffer.push(args);
-    }
-  }
-
+module.exports = function(trace, getStatsInterval, prefixesToWrap) {
   var peerconnectioncounter = 0;
   var isFirefox = !!window.mozRTCPeerConnection;
   var isEdge = !!window.RTCIceGatherer;
